Changing file structure. Fix models
