+++ KitMonitor/Server/KitMonitor.Server.csproj
@@ -27,6 +27,7 @@
 
   <ItemGroup>
     <Folder Include="Controllers\" />
+    <Folder Include="Migrations\" />
   </ItemGroup>
