+++ KitMonitor/Server/Models/Database/Entities/Company.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace KitMonitor.Server.Models.Database.Entities
+{
+    public class Company
+    {
+        [Key]
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public long Id { get; set; }
+        [Required]
+        public string Name { get; set; } = null!;
+
+        public IEnumerable<MonitorSystem> MonitorSystems { get; set; } = new List<MonitorSystem>();
+    }
+}
