+++ KitMonitor/Server/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": ""
+    "DefaultConnection": "Data Source=DESKTOP-DOH8HBN;Initial Catalog=KitMonitor;Integrated Security=True; TrustServerCertificate=True;"
   }
 }
